Insert first text
The text is added with screenshots.

An idea (by Gardinali) was just mark the parts where a image will be set and put them after. This will permit me to write in vscode and just put the screenshots on github. If this will increase my speed is for the future.